cotale-backend/pyproject.toml
[project]
name = "cotale-backend"
version = "0.1.0"
description = "CoTale TRPG 協作劇本編輯器後端"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "alembic>=1.16.1",
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "mypy>=1.16.0",
    "openai>=1.82.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "types-passlib>=1.7.7.20250601",
    "types-python-jose>=3.5.0.20250531",
    "uvicorn>=0.34.2",
    "websockets>=15.0.1",
]
